🚚 rename: MainMatch 컴포넌트를 Home의 components로 이동 #40
